requirements.txt
tensorflow==2.15.0
keras==2.15.0
flask==2.3.3
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3
matplotlib==3.8.2
scikit-learn==1.3.2
gunicorn==21.2.0 